Added to CameraController to fix mouse click not going through UI
