+++ Assets/02.Scripts/FSM/State/StateController/PlayerStateController.cs
@@ -273,6 +273,24 @@ public class PlayerStateController : MonoBehaviour, IStateController
         }
     }
 
+    public void SavePlayer()
+    {
+        Debug.Log("Button save going through");
+        SaveSystem.SavePlayer(this);
+    }
+
+    public void LoadPlayer()
+    {
+        PlayerStats data = SaveSystem.LoadPlayer();
+
+        Vector3 position;
+        position.x = data.position[0];
+        position.y = data.position[1];
+        position.z = data.position[2];
+
+        transform.position = position;
+    }
+
 }
