+++ Assets/02.Scripts/FSM/Stats/SaveSystem.cs
@@ -6,7 +6,7 @@ using System.Runtime.Serialization.Formatters.Binary;
 
 public static class SaveSystem
 {
-    public static void SavePlayer(SavePlayerLocation player)
+    public static void SavePlayer(PlayerStateController player)
     {
         BinaryFormatter formatter = new BinaryFormatter();
         string path = Application.persistentDataPath + "/playerSaveFile";
