+++ Assets/10.Plugins/NSubstitute.dll.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-
 guid: 0ffe43c04d6d8a84fb7168e497b4b922
-
 PluginImporter:
   externalObjects: {}
   serializedVersion: 2
@@ -14,26 +12,10 @@ PluginImporter:
   validateReferences: 1
   platformData:
   - first:
-
       Any: 
     second:
       enabled: 1
-
       settings: {}
-  - first:
-      Editor: Editor
-    second:
-
-      enabled: 0
-      settings:
-        DefaultValueInitialized: true
-
-  - first:
-      Windows Store Apps: WindowsStoreApps
-    second:
-      enabled: 0
-      settings:
-        CPU: AnyCPU
   userData: 
   assetBundleName: 
   assetBundleVariant:
